fixing some transition bugs
i put some events to go back to idle at the last frame.
for that the animation always needs to play until the end (some transitions allow it to blend before it ends).
a quick way to do this is to set the exit time and transition duration to something that equals 1 (for example 0.75 + 0.25).